requirements.txt
mlb-statsapi
pandas